requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Flask==2.2.5
Flask-Cors==3.0.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
pycparser==2.21
PyJWT==2.7.0
python-dotenv==0.21.1
requests==2.29.0
six==1.16.0
typing_extensions==4.6.2
Werkzeug==2.2.3
zipp==3.15.0
boto3~=1.26.129
